pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.1",
    "google-generativeai>=0.8.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
